+++ store-dto/src/main/java/hr/ivlahek/sample/store/client/product/ProductDto.java
@@ -1,10 +1,11 @@
 package hr.ivlahek.sample.store.client.product;
 
 import hr.ivlahek.sample.store.client.docs.DocumentationConstants;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -17,6 +18,6 @@ public class ProductDto {
     @ApiModelProperty(name = DocumentationConstants.PRODUCT_DESCRIPTION)
     private String description;
     @ApiModelProperty(name = DocumentationConstants.PRODUCT_PRICE)
-    private double price;
+    private BigDecimal price;
 
 }
